+++ Chapter1/Week06/MinMax2.java
@@ -1,8 +1,8 @@
 class MinMax2 {
   public static void main(final String[] args) {
-    int min = Integer.MAX_VALUE, max = Integer.MIN_VALUE;
+    long min = Long.MAX_VALUE, max = Long.MIN_VALUE;
     while (!StdIn.isEmpty()) {
-      int x = StdIn.readInt();
+      long x = StdIn.readInt();
       while (x <= 0) {
         System.err.println("All numbers must be positive, and thus " + x +
           " is not allowed -- please reenter:");
